Improve UI (#7)
* adds filter params
* adds how to use params to the README
* enables params from the ui & adds default query

+++ daemon/src/main.rs
@@ -2,10 +2,10 @@ use daemon::{
     create_folder, get_config_info, get_coordinates, get_forecasts, get_observations,
     save_forecasts, save_observations, send_parquet_files, setup_logger, Cli, RateLimiter,
 };
-use slog::{debug, error, Logger, info};
-use tokio::signal::ctrl_c;
+use slog::{debug, error, info, Logger};
 use std::{sync::Arc, time::Duration};
 use time::{format_description::well_known::Rfc3339, OffsetDateTime};
+use tokio::signal::ctrl_c;
 use tokio::sync::Mutex;
 use tokio::time::interval;
 
@@ -17,7 +17,7 @@ async fn main() -> Result<(), anyhow::Error> {
     // Max send 3 requests per 15 second to noaa
     let rate_limiter = Arc::new(Mutex::new(RateLimiter::new(
         cli.token_capacity.unwrap_or(3),
-        cli.refill_rate.unwrap_or(15.0 as f64),
+        cli.refill_rate.unwrap_or(15.0_f64),
     )));
 
     // Run once to start
@@ -35,7 +35,10 @@ async fn process_weather_data_hourly(
 ) {
     // defaults to once an hour
     let sleep_between_checks = cli.sleep_interval.unwrap_or(3600);
-    info!(logger, "wait time between data pulls: {} seconds", sleep_between_checks);
+    info!(
+        logger,
+        "wait time between data pulls: {} seconds", sleep_between_checks
+    );
     let mut check_channel_interval = interval(Duration::from_secs(sleep_between_checks));
     loop {
         tokio::select! {
@@ -98,6 +101,6 @@ async fn process_data(
         &root_path,
         format!("{}_{}", "observations", current_utc_time),
     );
-    send_parquet_files(&cli,logger, observation_parquet, forecast_parquet).await?;
+    send_parquet_files(&cli, logger, observation_parquet, forecast_parquet).await?;
     Ok(())
 }
